WIP Unfinished quadrature capability improvement
Adding axis-aligned quadratures, specifically a product set of
Gauss-Lobatto for the polar and Gauss-Chebyshev for the azimuthal.

Unfinished, primarily because of pervasive hard-coded logic using
a zero weight to denote starting directions instead of checking for
changes in eta (the polar angle after mapping) to denote a new level
in both Draco and libraries that use Draco.